+++ src/extension.ts
@@ -2,6 +2,7 @@ import * as vscode from 'vscode';
 import { createMacro, updateActiveEditor } from './commands/createMacro';
 import { createRepl } from './commands/createRepl';
 import { debugActiveEditor, debugMacro } from './commands/debugMacro';
+import { downloadAsset } from './commands/downloadAsset';
 import { openMacro } from './commands/openMacro';
 import { resetSharedContext } from './commands/resetContext';
 import { runActiveEditor, runMacro } from './commands/runMacro';
@@ -9,10 +10,11 @@ import { setupSourceDirectory } from './commands/setupSourceDirectory';
 import { showRunningMacros } from './commands/showRunningMacros';
 import { setContext } from './common/vscodeEx';
 import { MACRO_EXTENSION, MACRO_LANGUAGE } from './constants';
-import { MacroCodeLensProvider } from './language/macroCodeLensProvider';
-import { MACRO_TRIGGER_CHARACTERS, MacroOptionsCompletionProvider } from './language/macroOptionsCompletionProvider';
 import { MacroStatusBarItem } from './macroStatusBarItem';
 import { Manager } from './manager';
+import { DTSCodeActionProvider } from './providers/dtsCodeActionProvider';
+import { MacroCodeLensProvider } from './providers/macroCodeLensProvider';
+import { MACRO_TRIGGER_CHARACTERS, MacroOptionsCompletionProvider } from './providers/macroOptionsCompletionProvider';
 
 /**
  * Extension startup.
@@ -33,7 +35,9 @@ export async function activate(context: vscode.ExtensionContext) {
     { scheme: 'untitled', language: MACRO_LANGUAGE },
     { pattern: `**/*${MACRO_EXTENSION}` },
   ];
+
   context.subscriptions.push(
+    l.registerCodeActionsProvider(selector, new DTSCodeActionProvider()),
     l.registerCodeLensProvider(selector, new MacroCodeLensProvider()),
     l.registerCompletionItemProvider(selector, new MacroOptionsCompletionProvider(), ...MACRO_TRIGGER_CHARACTERS),
   );
@@ -43,6 +47,7 @@ export async function activate(context: vscode.ExtensionContext) {
     cr('macros.resetContext', (pathOrUri: string | vscode.Uri) => resetSharedContext(manager, pathOrUri)),
     cr('macros.debug', (pathOrUri?: string | vscode.Uri) => debugMacro(manager, pathOrUri)),
     cr('macros.debug.activeEditor', () => debugActiveEditor(manager)),
+    cr('macros.downloadAsset', (url: string, macroFile: vscode.Uri) => downloadAsset(url, macroFile)),
     cr('macros.new.macro', (content?: string) => createMacro(context, content)),
     cr('macros.new.macro.activeEditor', () => updateActiveEditor(context)),
     cr('macros.new.macro.repl', () => createRepl(context)),